Correct CDATA handling in adapter
By default, CDATA content is escaped by the XML parser at parse time
(e.g. '>' is replaced with '&gt;').  With the ElementTree and
cElementTree parsers, it was necessary to hack internals to work around
this issue.  lxml provides an option to the parser which causes it to
leave CDATA segments intact.

Adding unit tests to ensure this behavior is not regressed again.

Change-Id: Id2d6de8827531c330af5b6172128d70035b7a65f