Utilize Element Lists
The wrappers need the ability to form complex types.  This change set
provides the ability for the wrappers to support complex list types that
update the backing element tree.

Also includes a list type (useful for properties or simple lists) that
will call an action on set.

Starts the conversion of complex properties over to types that support
this new model.

Change-Id: I106cba03cefec113abfbe86423c384f73c252b7d