Add constants for extended attribute groups
Extended attribute groups can be provided per the 'xag' variable on the
read operation within the adapter.  They are mostly used for the Virtual
I/O Server.  They provide reads for specific attributes, and can greatly
reduce the overall read time of the element.

This change set puts a set of the common attribute groups in the Virtual
I/O Server wrapper.

Change-Id: I38d450313c8be6bd8f32967ed2f08529d0e0ddc0