Support MTMS query in Managed System
Multiple projects that build on pypowervm need to read a query of the
managed systems and find the appropriate managed system based off MTMS
and Serial Number.

This change provides a function in the wrapper to query for this.  Also
updates the MTMS format to Model-Type*Serial.

Changes the equality method to remove attribute checks.  This is due to
name spaces overlapping.  Another change set should look at checking
attributes in that equality statement.

Change-Id: Id51e4770f264e69e20105c0dbacbd5bec7ed32bf