Add function to find a wrapper from a list
This change adds a utility function that enables users to find a given
wrapper from a list of wrappers based on a UUID.  This is helpful when a
feed has been loaded, but they need to find a specific UUID.

Typically users would load just the element instead of the entire feed,
but there are rare scenarios where they need the whole feed and a given
element.  This change really applies to those less common scenarios.

Change-Id: Iba1a383be450188d402bcff7d49859156bbd3ec3