Sync load groups back to Network Bridge
The main way to interact with a Network Bridge to distribute VLANs
across the load groups is via the Virtual Networks on the Load Group.
However, the NetworkBridge also has a list of all of the Virtual
Networks.

Modifications (updates) to the Network Bridge require that the Virtual
Networks in both elements are 'in sync' (though less resolution on the
Network Bridge since it is simply a list of all Virtual Networks).

This change set will make it so that the Network Bridge's Virtual
Networks are updated to stay in sync whenever they change on the Load
Group.  It adds a callback mechanism on update calls, as well as extra
data being passed in to the elements via kwargs.

Change-Id: I21ee00136df4ac3662b768eb7504c0b15fcf3533