Fix VolumeGroup capacity numbers
VolumeGroup capacity was previously
a) Incorrectly documented as represented in MB (it is really GB).
b) Rounding via int().

This change set fixes the docstrings and changes the conversion to
float().

In the process, also eliminated entry_wrapper.Wrapper's
get_parm_value_int() method in favor of adding a converter=<callable>
param to the generic get_parm_value() method.

Change-Id: I1955a7cfd85cdc410bcc7511b6671dd862f5250f