Clean up of upload_lv components
The upload_lv component has been modified to support disjoining the file
and disk size on the upload of vDisks.  This allows for bigger disks to
be defined than the image uploaded to it.

Also removed a significant amount of return parameters that, since the
clean up method was introduced.  Now it returns the file_uuid if the
cleanup failed.  This also removes the scenario where a file_uuid that
was deleted via the cleanup was returned (invalid return data).

Change-Id: I9c5274a0d7b1feb8a244a5145cf8b1cac672843e