Refactor get_parm_value_bool()
Changed get_parm_value_bool() to use the generic get_parm_value() with
a new custom converter.

This turned out to be a much smaller change than I was expecting.  I
had thought that get_parm_value_bool() was still returning a string
value ('true' or 'false') and had a goofy default ('*unset*'), but on
closer inspection, that was not the case.  I also thought I would be
able to get rid of the method entirely and replace calls with
get_parm_value() with the new converter util; but that would have
changed the 'element not found' behavior from False to None.

Change-Id: Ib850480aec502b26a3713df9b8a1ae3ebda727dc