Provide I/O hardware getters
The wrappers need to be able to expose more of the underlying hardware.
Specifically the I/O configuration.

This change set provides initial getter methods for this.

Specifically the following is done:
 - Added the PartitionIOConfiguration
 - Added the IOSlot
 - Added the IOAdapter
 - Modified the PhysFCAdapter
 - Added the PhysFCPort

While the backing schema is more verbose than this, the API hides some
of the additional complexity (ex. AssociatedIOSlot and
RelatedIOAdapter).

This change set only approaches the getters at this point in time.
Future change sets will determine the correct setters required.

Change-Id: Ida1e4c796808637584a2fc2e79a7e43d9518383d