Provide hdisk discovery utility
When connecting a disk via standard Fibre Channel methods, the disk then
needs to be discovered by the host system.  The utility to do this
quickly is the LUA recovery API.

This change set provides a discover_hdisk method (which behind the
scenes calls the lua_recovery).  This will find the appropriate hdisk
for the invocation.

Change-Id: I25f33c667ca0f8780bcdd2f7863ad99667eb33b4