Provide mappings for PV and LU
The vSCSI mapping creations extend to PV (for classic vSCSI mappings)
and to LU (for SSP connections).  This change set provide
implementations for this.

Given that the LU requires more than the disk name to be passed in, this
method changed the way that the bld method works by passing in the
pre-built child element.

Provides additional unit tests for these new types.

Change-Id: I693153c6f703093b22ecd1fc4fdbd05b077981d2