Maintainability of VSCSIMapping.backing_storage
Previously, VSCSIMapping.backing_storage was detecting the type of
backing storage by looping through a hardcoded list of know storage
element types.  Thus, if a new storage type was enabled, it would also
have to be added to this loop.

This change set eliminates the hardcoding and takes advantage of
ElementWrapper.wrap()'s ability to determine and return the appropriate
wrapper subclass based on the registry of known types.

Also fixed some docstring errors/typos.

Change-Id: Icfea11669150dd66925225dfb2b8f6323982fd79