Provide WWPN generation
Some tools will need to generate the WWPNs for client VirtualFCAdapters.
This is usually done during the Adapter creation.  However some
scenarios may want to generate this ahead of time.

This change set provides an API to do this.  Currently it provides a
random generation method.  It is planned to support a globally unique
WWPN naming scheme in a future change set.

The liklihood of WWPN overlap is 1 in 281 trillion.  Within a given SAN
fabric the liklihood of overlap is virtually non-existant.  Nonetheless,
the future change set should resolve the remaining probability of
overlap.

This change set has a minor update to close_vterm to remove a 'no new
line at end of file' warning.

Change-Id: I144556ab145268eab16eb7dc0effa104f8bfe3fe