Adapter.refresh(entry_wrapper)
Implement Adapter.refresh(entry_wrapper): performs a GET of the entity
represented by entry_wrapper and handles 304 (Not Modified) by
returning the original wrapper.  This provides a consistent mnemonic
for consumers to retrieve the latest version of an object without being
aware of the low-level semantics of etag and HTTP response codes.

Change-Id: Idd5708fc85e90fcd143f57357ce0b0db0f62a956