Add lpar builder and standardizer
Added the lpar builder class to build an lpar wrapper from a
dict of resource settings.
Added a standardizer that fills in missing resource values.
The standardizer also performs the bulk of the validation
of the lpar attributes.

Added field classes to help with validating common forms of
the lpar attributes.

Change-Id: I614fdff45b7b95a75867ea87e2211439284f4fcb