EntryWrapper.update takes xag, returns new wrapper
Previous change set to enable update on a wrapper omitted a couple of
points, remedied by this change set:

o Extended attributes are (currently) not returned in the SELF link on a
GET, so we have to account for them by adding a parameter to the new
update method.

o The method previously returned the Response from Adapter.update.  It
now returns the updated EntryWrapper.

Change-Id: I582cb6dc3694538db6129b82bfdaf46089d3b21e