PoC: exploit implicit child ordering for Node
This change set enables Wrapper subclasses to take advantage of the new
adapter.Element.inject method to allow assignment of child elements in
any order.

This change set proves the concept with the Node ElementWrapper by
reordering the assignments in its bld() method.  Existing unit test
test_cluster_ssp explicitly tests the ordering of the XML output.  Node
is rich for this test because it uses set_parm_value, _find_and_seed,
and set_href.

It is possible that other entry_wrapper changes will be found necessary
as this change is further integrated.

Change-Id: Ie89631e632d9717e097515f7b7697ed476941874