SCSI Mapping Logic in pypowervm
The vscsi mappings are the virtualized SCSI drivers that link together a
device on the Virtual I/O Server and the client VM.  There is the SCSI
Bus (often referred to as the mapping) and the SCSI devices.  A single
bus can support many devices.

The clients of pypowervm typically just want to indicate that they need
a device added to the VM.  They typically do not want to deal with the
hastle of determining if a mapping should be fused to another mapping.

This change set provides functions to add mappings to a given VM and
allowing the logic of determining which bus to put that mapping on to
the task.

The upload function received a small change, so that invokers could use
the vdisk returned from the upload immediately within the mapping
function.

This is a DISRUPTIVE change.  There was also a bug where the mappings
presented to the clients only allowed one device.  This change fixes
that issue.

Change-Id: I370f4dd755d42c76219cc18af66ebb24cfef9f09