upload_new_vdisk/vopt: crt_vdisk/_upload_stream
In anticipation of needing to share code for uploading to LUs, this
change set refactors pypowervm.tasks.storage.upload_new_vdisk and
upload_vopt.

o The code to create the new VDisk is extracted from upload_new_vdisk
into the new method crt_vdisk.  This is a direct corollary to the
existing crt_lu.

o The code to upload a data stream to a VIOS File and clean up its
metadata was common in upload_new_vdisk and upload_vopt; and will be
common in upload_new_lu when it is written.  This code, including the
one-line helper upload_cleanup, is extracted/combined into the new
method _upload_stream.

The only functional change is the return value from the two upload_*
methods, which is now the whole File EntryWrapper as opposed to just its
UUID.

Change-Id: I339e59371e22b0eb02b54d4a7c3a8136d88cd4d7