Move the storage adapters to the correct location
The storage adapters (client/server, scsi/fc) had been stored in the
virtual_io_server.py as they were components of the mapping.

However, they are independent objects that get used as part of the
mapping.  They can also exist independently.

As such, their location more correctly is in storage.py.  This change
migrates the components to that location.

Change-Id: Ibec9ee865ded785ee25e0b2034523d26602a0ecb