Provide WWPN mapper
The Virtual FC Adapters need to be 'mapped' to the physical ports that
the traffic will flow through.  Determining the mapping can be
difficult.  Typically the desire is to map across VIOSes, and then
balance across the ports in the VIOS.

This change set provides a utility in the wwpn.py that provides a
mapping back to an invoker.  This takes into account many best practices
when mapping across VIOSes and ports.

The method does not actually perform the mapping, but instead just gives
the user the recommendation of how to map the ports.

Change-Id: Iab42c763528018c9502948310c9c0b2a5efc5bf5