Fix adapter read when not using caching
In the Adapter's read_by_path method, the _locked_refresh method
was always being used whether caching was requested or not. The
effect was that in cases where reads of the same path happened
within the refresh interval, it tried to access the cache,
which was not setup and thus caused an exception.

This change avoids performing the read operation through the
_locked_refresh method when caching is not requested or
the path is not cacheable.

Change-Id: I71444cb3f1dcc887e07e0410b36bb0491d36a4dd