Include VIO mapping functions
The derive_npiv_map function provided a utility to derive which NPIV
virtual ports should be mapped to a given physical port.  However, it
introduced a new data structure that was unique, and didn't necessarily
provide the actual mapping.

This change set introduces two new functions to make use of that new
data structure.  These are the:
 - add_npiv_port_mappings: takes in the port mapping list and will apply
   them to the necessary Virtual I/O Servers.
 - remove_npiv_port_mappings: takes in the port mapping list and removes
   them from the necessary Virtual I/O Servers.

This logic has intelligence built in so that it does not call updates on
Virtual I/O Servers that do not require it.

A side effect change was in the virtual client wwpns.  They are lower
case, which is non-standard to the rest of the APIs wwpns.  The wrapper
was updated to hide that so that users can use a consistent format for
WWPNs

Change-Id: Iebe65989fc47d8b034c6813200252c328775b6eb