Improvements to get_req_path_uuid
UUIDs in the PowerVM REST API are case sensitive.  In particular, a
web/File object must refer to the target VIOS using its UUID in
the original UPPERCASE; else the PUT of the File/contents will result in
a 400 error with the message:

TargetVirtualIOServerUUID is not found or does not represent a VIOS.

This is a problem in our cluster.Node.vios_uuid getter, which was using
get_req_path_uuid, which by default "sanitizes" the UUID by converting
it to lowercase.

This change set adds the option to get_req_path_uuid to preserve the
original case.  Also adds logic to ignore query parameters (?foo=bar)
and fragments (#baz) if present.

Then we use the new option in cluster.Node.vios_uuid to make sure the
UUID comes back in the correct case.

Change-Id: I0584fdddae3d249451264e1eb98b7a0c439bf3ee