Round LU size up to 2dp
LU size was being rounded to the nearest... something by the platform.
(The observed example was 4.3MB being rounded to 4.0MB.)  In cases where
the LU is being created just big enough for the image, this results in
ENOSPC errors attempting to upload the image.

This change set adds a method to round a GB size *up* to a level
suitable for the platform.  We chose a default of 2dp (approx. next
highest 10MB) as a good trade-off between precision and space
conservation.

Change-Id: Iaa174bf6ae651015be07a8559199785bebfb9e9f