Create the vSwitch on the client CNA create
When users create a Client Network Adapter, they have to specify a
virtual switch by name.  However, if that virtual switch did not exist,
the code previously threw an error.

This change provides the ability to auto create the vSwitch if it is not
found.  This saves the user a couple of extra steps in their invocation.

Change-Id: I0238dd05caa8c1435e3010b3d2f5157fa2fcc55d