Fallback search by looping through GET feed
Previously, EntryWrapper.search only supported searching via REST API
/search URIs.  This meant that a given search key had to be a) supported
by a /search URI, and b) entered into the EntryWrapper subclass's
search_keys dictionary.

Consequently, we have a pervasive pattern in the code that performs a
'manual' search by GETting the entire feed of the relevant object and
looping through it looking for a particular field/value.

This change set is a step toward eliminating that pattern and allowing
EntryWrapper.search to be used anywhere such a search is needed.

The change herein encapsulates the GET-feed-and-loop logic into the
EntryWrapper.search method as a fallback when the subclass doesn't have
search_keys at all, or when the desired key doesn't exist in
search_keys.

Subsequent change sets will expand this method to support CHILD types
and extended attribute groups.

Change-Id: Ifecb9e0d9fcce09f0013077e3618270c425dc4d7