Fix the SCSI mapping function
The SCSI fuser had an incorrect assumption that a given mapping could
have multiple elements.  Further inspection of the schema indicates that
a mapping will only have a single storage element.

A SCSI Bus (which consists of a SCSI Server Adapter and Client Adapter)
may have multiple storage devices.  But the SCSI Mapping indicates a
unique pair of adapters and storage device.

Therefore if a SCSI Bus has multiple storage devices (VOpt, VDisk,
etc...) then there will be a mapping for each device.

The SCSI Mapper has been updated to reflect this for both the add and
remove functions.

Also includes a fix for media_name being set to none to properly remove
the media mapping.

Change-Id: I84002d5a8e0472743ef048f09a011e86089e6be2