Return the vOpt on the upload
The upload vOpt method worked differently than the others as it would
simply return a possible file uuid.  The other methods would actually
return the element that it was trying to upload.  This was inflexible
and required users to query to find it.

This change set provides the vOpt post upload.  The algorithm may be
expanded later to support faster query.

Change-Id: Ia5b8d6360bde7f5942ba7429b1d000809c64d716