Update related href build to not include groups
Due to the recent optimization to specify group=None on most requests,
the related href build method was inadvertantly affected.  Links as such
should not have extended attributes.

This change fixes the issue and adds a test case to make sure that
future URI optimizations we make won't impact it again.

Change-Id: I75ffee55a3c7fa675ad57f5f67bf4cf4ecc0b3c1