Initial development of is_hmc flag
The PowerVM REST API may be backed by an HMC or a next generation API
server.  The goal of pypowervm is to be agnostic at a high level between
these two.  However, there are scenarios where there is slight variation
between the two services.

This change set introduces an is_hmc flag.  It is brought into the
Session.  Right now it is hard coded.  A future change set will
dynamically detect the flag upon the login of the session.

The adapter has a wrapper property around the session for easier access.

Finally, this change set adds a concept of only calling the
VirtualNetwork creation if on an HMC.  The next generation API only
realizes VirtualNetworks based on the VLAN layout.  It does not allow
for explicit creation (outside of the adapters).

Change-Id: I369395abfc5e195eeb32a30f0b03a7e51758b0cb