Remaining support for happy path NetBridge bld()
Added:
    - BackingDeviceChoice/EthernetBackingDevice
        - AdapterID/DeviceName
    - vswitch_href argument
    - VirtualSlotNumber
    - Tests for the above

Notes:
Some of these will probably not be here permanently. This was what
was necessary to get NetBridge creation working against neo given
the current state of everything.

BackingDeviceChoice/EthernetBackingDevice classes may go away if
they aren't deemed necessary. Added them so we can extend as we
feel we need to.

Change-Id: Ifb8802ebe151819bff8bc40df80ab4ca0457c7e9