Support related_href property
With the new function in place to default group=None, the URIs from
tje API come back with a link that includes group=None.  That is
expected.  However, the unintended side effect of this is that the href
element is used often between elements when building relationships
between entries.  This is often done in the network code for instance.

Relationships require the base URI to be specified (no group= elements,
just the core URI).

This change set introduces a 'related_href' property.  This builds on
top of the existing href property, but uses the dice feature to cut out
the extra data that may be contained.  Therefore the response to this
method should return something usable for relationships between
elements.

This change set also updates test data and the networking files (which
appear to be the only ones in pypowervm using relationships currently).

Change-Id: Ib143557576e52718f6977a2bcf67a978fb0b92a7