Initial API Traits configuration
This change set provides the initial framework for API traits.  These
are aspects of interacting with the API that can change depending on the
backing PowerVM REST Server.

A future series of change sets will make this far more robust, adding
additional attributes.

This first change set provides a flag for the virutal network
utilization on certain operations.  The next set of changes will start
to inspect the session to determine the appropriate value for this.
Right now it is hard coded to false, to reflect the next generation API.

This also updates the cna.py to make use of the new attribute.

Change-Id: Id6e507ba2fa99eb9ca3496e634004fed7ad14523