Remove pypowervm.wrappers.constants
In keeping with the move toward putting constants in the files in which
they're used, and making them as private as possible, this change set
moves all remaining constants out of pypowervm.wrappers.constants and
eliminates that module entirely.  Almost all constants used by multiple
modules were placed in the module in which they made the most sense.
The remaining couple of widely-used orphans were put into
pypowervm.const.

Change-Id: I1de61c180153d5a18016b010e07cb06b7ba8fce0