Bug fixes for wwpn API
The wwpn.py file had a few bugs in it.  Since the switch to group=None,
the API needed to be updated to include the extended attribute for
STORAGE.  It also was found that in some versions of the API, the WWPNs
coming back from the API were a different case than expected.  This
update includes broader sanitization of the WWPNs, even for the data
passed back from the API itself.

Change-Id: I2d7cf17bf47ed8293fa09c84edb550d2ccc90d88