Support modification in wrappers for VLANs
The vnet_aware trait was recently added.  This trait indicates that
modification of the VLANs should be executed against the TrunkAdapters
within the Network Bridge rather than via the VirtualNetwork List.

This change set provides base level infrastructure within the wrapper to
allow for that modification.  It allows SEAs to accept new TrunkAdapters
and for the addl_vlans within a TrunkAdapter to be modified.

Additionally, given the shorter names of the wraper classes now, this
change set updates the virtual_network_uri_list to a more concise name
of vnet_uri_list.  That brings it more in line with the wrapper class
name that the element now has.

Change-Id: I267669ebb07ee6d329bc2c615dbbc2dee4627a41