Fix Accept header for quick prop with query params
Adding query parameters (say, ?group=None) or a #fragment to a quick URI
doesn't actually affect the GET call, which safely ignores them.
However, these URI add-ons exposed a bug in the code that determines the
proper Accept header for the GET request, resulting in a 406 (Not
Acceptable) for URIs of the form
/rest/api/{servlet}/{object}/{uuid}/quick?group=None

This change set ensures that query parameters and fragments are ignored
when calculating the Accept header, and also expands unit test coverage
for the code path in question.

Change-Id: I9c930903baaf7d078d2cd06ffe885599a547822d