Provide initial refactor of network bridger
The network bridger needs to be refactored to support two flows.  The
existing that works against the VNet.  A second version that bridges the
VLANs directly on to the Trunk Adapters.

This change set provides the refactoring that abstracts the two flows
out.  There is now a common class that does that majority of the work.
Then two subclasses, one for VNet modification.  A second class has been
stubbed out (but not implemented) for the VLAN modification.

A subsequent change set will provide the VLAN implementation.  For now
the methods in the stubbed class 'pass' so that users of non-VNet aware
APIs do not have exceptions thrown when called.

This change set also adds a fixtures.py class to make testing and
stubbing out the adapter easier.

Change-Id: Idb1a2c1e7bca5efde8b04548b84fd6f80ff3cdea