Add trait for colocation to PowerVM API Server
In certain scenarios, the pypowervm libraries may be running on the same
system as the PowerVM API.  During that, certain optimizations may be
enhanced.  Specifically, uploading to the API a file can be optimized
dramatically.

This change set adds a trait to the library which allows us to determine
if the code is running on the same system.  It is a coarse check,
flagged by the file auth flag.  Any user based authentication is not
checked for.  Because of the branches of complexity that introduces,
that won't be checked until there is demand for it.

Change-Id: I49cb4cb04c2357eb28f529a16d050d89eb490d17