Add VLAN path for network bridger
The network bridger needs to support multiple paths depending on the
PowerVM API.  One that works against Load Groups.  Another that works
against the Trunk Adapters directly.

This change set provides the functions to operate against the Trunk
Adapters directly.

This modification implements the NetworkBridgerTA.  Uses the traits in
the wrapper to determine how to derive the VLANs. If not using virtual
networks, will pull from the Shared Ethernet Adapters directly.

Change-Id: I55b47528114af527c9ef8df30d8fd808b05610ef