Add VLAN path for network bridger
The network bridger needs to support multiple paths depending on the
PowerVM API.  One that works against Load Groups.  Another that works
against the Trunk Adapters directly.

This change set provides the functions to operate against the Trunk
Adapters directly.

This modification implements the NetworkBridgerTA.  In addition, also
makes some changes to the base wrapper class to take in flags to
determine how to drive functions such as listing VLANs.

That modification was required because as the Trunk Adapters change, the
base way we list VLANs can not be derived solely off of the Load Groups
(it is stale data from the Get).

Change-Id: I55b47528114af527c9ef8df30d8fd808b05610ef