Support new file upload mechanism
The REST API server supports a new 'local disk' file upload mechanism.
If you're running on the same system as the REST API, instead of sending
a file to the REST API for upload, one can instead simply write to a
file pipe.

This reduces the overhead on the REST API server dramatically (lower
CPU, faster performance of the upload).

This change set adds support for this flow.  Specific changes:
 - Add a new asset_file property to the vios_file wrapper.  This only
   shows up if using the new upload
 - Change the enumeration types to be more terse.
 - Add a new vio file enumeration type
 - Update the file upload for LU and vDisk to support this new
   mechanism.

Change-Id: Icf8d91c9aaa242e34317a85c4ca31ddcde37a248