Fix the update wrapper for duplicate groups
The recent updates to the path code resulted in duplicate groups being
specified.  The check to determine if the group had already been
specified was removed.

This adds a check specifically to the update aspect of the wrapper.  If
the URI already has a group, it should be used as that is the most
accurate.  If it does not, the path should be extended.

Change-Id: Iec00a75ff4204c5d830f969bbb05559dac8f8b79