Embed traits in entities.Entry and .Element
APITraits, recently introduced to facilitate cross-version compatibility
of pypowervm with different REST API servers, are sometimes needed
within Wrapper instances to make decisions as to how certain getters and
setters should be implemented.  Because APITraits are linked to the
Session, the most effective way to trickle them down to Wrapper
instances (without requiring the consumer to pass them into the Wrapper
methods themselves) is via the following paths:

Response => Feed => Entry => EntryWrapper
         => Entry => EntryWrapper
         => Element => ElementWrapper

This change set creates the framework and implementation such that a
given ElementWrapper or EntryWrapper instance has a .traits property it
can consult to make API-specific decisions.

Change-Id: I63723a51f229786a42eb9d8193a4516d8d0da15b