Incorporate X-MC-Type into Session and APITraits
Newer versions of the PowerVM REST API include the X-MC-Type header in
all HTTP responses to indicate whether the underlying management console
is 'HMC' or 'PVM'.  This change set reads that header and incorporates
it into the previously-stubbed-out logic backing several of the traits
in APITraits.

If the X-MC-Type header is absent, 'HMC' is assumed.

Change-Id: Ieddd069545c9a91bb47317d7bd97eaa3a6c4d916