Fix additional VLANs bug in network_bridger
This change set fixes a couple of issues in the network_bridger.
The first is a bug was that 'all' of the VLANs were not taken into
account when deriving the arbitrary VID.  This led to a TrunkAdapter
having a PVID that matched an additional VLAN ID.  The fix is to pass in
all of the VLANs when deriving the new arbitrary VID.

The second fix was to make sure that all of the vlan IDs passed in were
converted to integers.  There may be some consumers that have VLAN IDs
as strings.  This resolves that issue.

Change-Id: Iff2bea74ed1909dcf794e59458f3428ea55ca93c