WIP: Ensure SELF links exist on all Entrys
This version started down the path of trying to pass hrefs into
unmarshal_atom_entry so that, if the entry was missing a SELF link, we
could generate it from that param.  The problem is that this only works
for UOM-style URIs, which always have proper SELF links anyway.  It
doesn't work for PCM URIs, which don't follow the same pattern.  I'm
abandoning this line of thought for now, but forking a separate change
set to preserve the Atom base class, Feed links, and extend_basepath.

====

I'm at the point where I need to figure out how to pass the request URI
into unmarshal_atom_entry().  This leads to wanting the request URI (not
just reqpath, which lacks scheme://netloc) in the Response.

Also need to decide whether to augment unmarshal_atom_feed() similarly
to ensure SELF links on that guy if it doesn't have them.

Change-Id: I43df8fef58f04d24768f181328ff88de7ead3a53