Add 'active' pfc ports in vios wrapper
This changeset provides the function to determine the active ports for a
physical FC adapter within a VIOS wrapper.

The logic is currently very coarse.  It is anticipated that the PowerVM
API will provide better functionality to determine this in the future.
However, this is a first pass that works for the majority of scenarios.

Change-Id: I9d4088efb3bb1efe3d638d23ebb8103a810168c8