Implement EntryWrapper.create()
With the advent of adapters reliably embedded in all EntryWrappers, we
can now implement a create() method that can simplify and replace
existing calls to adapter.create().  The EntryWrapper.create method
needs no arguments for root types; and a parent type and UUID for child
types.  The parent type may be specified as a class (e.g. net.NetBridge)
or a string schema type (e.g. "NetworkBridge").

Change-Id: I5060718cc6c8f277f6c707ceca3ac56bfdc34fa8