Fix close issue if no tty
The close of the terminal should not throw errors.  Typically, the close
of the tty will be called in addition to deleting an LPAR.  However, if
the terminal was never opened, then the close should no op.

This change fixes an issue in that behavior where the tty was not found,
thus causing an issue on close.

Change-Id: Iadbd182d28e25c26df7c74714b6e7b0b072a8691