Util method to generate sanitized file name based
Based on PowerVM REST API's FileName.Pattern, the new method
pypowervm.util.sanitize_file_name_for_api accepts a name and an optional
prefix and suffix and returns the assembled name with any illegal
characters (according to FileName.Pattern) replaced with underscores;
and at a maximum length of 79 characters (also per FileName.Pattern).
The base name is truncated such that the resulting string fits.  The
prefix and suffix are never truncated, but ValueError is raised if they
are too long to allow any of the base name to remain in the string.

Change-Id: Ia5bc9f35ee9b657be6d8194df516053826de4975