Limit storage upload to three parallel operations
This change set limits the number of concurrent upload operations to 3.
This is because the backing Virtual I/O Server has some issues moving
past three.  Limiting increases the resiliance of the code, and
ultimately leads to the best performance overall.

This change set also adds some test coverage around the adapter upload
in general.  However, the upload semaphore shouldn't go in the adapter's
upload_file because of the coordinated disk upload.  Given that method,
we need the upload to work at the higher level.

Change-Id: Ibd95050402db1a9c39ac0a116a9ec6f4a7b15bef