Util method to sanitize partition name
The new method pypowervm.util.sanitize_partition_name_for_api accepts a
name and returns it with any illegal characters replaced with
underscores; and at a maximum length of 31 characters.  An optional
parameter indicates that truncation is not allowed, resulting in
ValueError if the input parameter is longer than 31 characters.

Change-Id: I1997cf3bb2ee8e4d54604f5f9174af72463267db