scsi_mapper.find_maps, gen_match_func
Factor out and publicize pypowervm.tasks.scsi_mapper.find_maps and
.gen_match_func.

find_maps filters a list of SCSI mappings according to an input client
LPAR ID and a function to match the backing storage element of a
mapping.  This functionality was formerly embedded in the
_remove_storage_elem helper.

gen_match_func generates a matching function suitable for passing to
find_maps.  This was formerly the private _remove_search_func helper.

Change-Id: Ife774d8ce928ad63597a572bd89f0a9ca62a554f