Support PCM Data Sources
This change set introduces the ability to parse through two types of
PCM Long Term Metrics.  They are phyp and VIOS.

The PHYP metrics represent the data sources from the hypervisor itself.
Certain pieces of data come back here that need to be correlated to the
Virtual I/O Server metrics (which contain other details).

These two are typically tied together to form a complete picture of the
metrics within the system.

This first change set provides the 'raw data'.  A subsequent set of
change sets will provide additional features such as finding the latest
phyp and vios elements from the feeds, determining 'system' metrics,
etc...

But before those can be built up, the raw data is needed which this
provides.

Also includes tasks to deal with updating the LTM preferences and
querying the data as it is non-standard.

Change-Id: Iec347296b5762932377cc40f2592b119ebb1de2d