Reconcile metrics to per VM basis
This change set provides a task and set of new classes that collapse the
metrics from the existing PHYP and VIOS level down to a per-VM level.
This consolidation should make it far easier for consumers to determine
metrics about a given LPAR.

The new metrics use the prefix 'Lpar' in front of them and are contained
within the task class instead of wrappers package.  This was done as
they are not directly provided by the API.

Change-Id: I7c6062926ff25cbbfb54f79595ab79f26820a366