SCSI ops can take VIOS wrapper or UUID
Some callers of pypowervm.tasks.scsi_mapper.add_vscsi_mapping and
.remove_*_mapping may already have retrieved the VIOS wrapper with the
SCSI_MAPPINGS xag.  In the previous code, they were forced to pass in
only the VIOS UUID, and the scsi_mapper methods would perform another
lookup of the same data.  With this change set, the parameter may be a
UUID *or* an existing VIOS EntryWrapper.  In either case, the up-to-date
VIOS wrapper (with SCSI mappings) is returned.

Change-Id: I842758a0e51222890292952d1e70ce49a658ac21