Add LPAR Metric Cache
WIP: Unit Tests being developed

Obtaining the LPAR metrics is expensive.  Doing so for each LPAR is too
costly at the API level for the user.

This change set provides an LPAR cache that gives access to the latest
metrics (as well as the previous metrics).  It simplifies the access of
the metrics down to just requiring the pypowervm adapter, host_uuid and
the lpar_uuid.

The cache also updates its data sources as needed upon subsequent calls
into it.  This allows for users to not need to worry about when the
cache was updated.  This helps simplify the interaction with pcm
dramatically.

Change-Id: I2ec2aeb5828d5fb90cc9f564e32e2c1b17c4f07f