Enable sonar scans through tox
Enable running sonar scans on the source. In order to run the
tox sonar environment you must have created a host mapping
in /etc/hosts for the name 'sonar-server'. If the sonar
server were on the local host then the entry might be:

127.0.0.1  sonar-server

The sonar-runner command must be installed locally.  See:
http://www.sonarqube.org/downloads/

The following environment variable must be set in order
to log onto the sonar server:
  SONAR_USER
  SONAR_PASSWORD

An example invocation:
  SONAR_USER=user SONAR_PASSWORD=password tox -e sonar

Sonar output is placed in:
  .sonar/

Change-Id: Ia06be9d6a513716efe74debc13fd0febc0b78e27