Enable sonar scans through tox
Enable running sonar scans on the source. In order to run the
tox sonar environment you must have created a host mapping
in /etc/hosts for the name 'sonar-server'. If the sonar
server were on the local host then the entry might be:

127.0.0.1  sonar-server

The sonar-runner command must be installed locally.  See:
http://www.sonarqube.org/downloads/

Change-Id: Ia06be9d6a513716efe74debc13fd0febc0b78e27