Update monitors to read_by_path
The PCM monitors were originally reading by href.  That method optimizes
the URI to include the ?group=None optimization to the API.  That flag
does not apply in the PCM context.

This change set adds a parameter option to the read_by_href to
optionally take in the xags.  This works similarly to the other methods
in the adapter now.  If the user wishes to not have a xag added, they
specify xag=[].

This change set also fixes a bug in the PCM that uses this code.  It
gets the response body from the json data.

Change-Id: If7182c0aec3c716162aa9a52cd948cd43a8db7af