Update monitors to read_by_path
WIP: Needs to handle the read in vm_metrics

The PCM monitors were originally reading by href.  That method optimizes
the URI to include the ?group=None optimization to the API.  That flag
does not apply in the PCM context.

This change set switches the reads to use read_by_path, which will
respect the original URI in a non-optimized path.  This allows the reads
to flow through properly.

Change-Id: If7182c0aec3c716162aa9a52cd948cd43a8db7af